@@ -9,15 +9,20 @@ import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.net.Uri;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.v4.app.Fragment;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.CardView;
 import android.util.Base64;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Toast;
 
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
@@ -31,9 +36,13 @@ import java.io.InputStream;
 
 import de.hdodenhof.circleimageview.CircleImageView;
 import razon.lostandfound.R;
+import razon.lostandfound.activity.AdminActivity;
 import razon.lostandfound.activity.HomeActivity;
 import razon.lostandfound.activity.LoginActivity;
+import razon.lostandfound.activity.MainActivity;
+import razon.lostandfound.model.UserGeneralInfo;
 import razon.lostandfound.utils.FirebaseEndPoint;
+import razon.lostandfound.utils.FragmentNode;
 import razon.lostandfound.utils.MyTextView;
 import razon.lostandfound.utils.SharePreferenceSingleton;
 import razon.lostandfound.utils.SimpleActivityTransition;
@@ -56,6 +65,7 @@ public class ProfileFragment extends Fragment {
     private MyTextView designation;
     private CardView changePhoto;
     private CardView logout;
+    private CardView update;
 
     private static final int RESULT_CANCELED = 0;
     private int GALLERY = 1, CAMERA = 2;
@@ -67,6 +77,8 @@ public class ProfileFragment extends Fragment {
     DatabaseReference reference;
     ProgressDialog progressDialog;
 
+    String usernameAdmin;
+
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
@@ -74,6 +86,17 @@ public class ProfileFragment extends Fragment {
         // Inflate the layout for this fragment
         View view = inflater.inflate(R.layout.fragment_profile, container, false);
         initView(view);
+        usernameAdmin = SharePreferenceSingleton.getInstance(getActivity()).getString("username");
+
+        if (usernameAdmin.equals("admin12")) {
+            worksOnAdmin(view);
+            logout.setVisibility(View.GONE);
+            update.setVisibility(View.GONE);
+        }else {
+            logout.setVisibility(View.VISIBLE);
+            update.setVisibility(View.VISIBLE);
+        }
+
         populateData();
 
         changePhoto.setOnClickListener(new View.OnClickListener() {
@@ -89,9 +112,9 @@ public class ProfileFragment extends Fragment {
             @Override
             public void onClick(View view) {
 
-                SharePreferenceSingleton.getInstance(getActivity()).saveString("user","0");
-                SharePreferenceSingleton.getInstance(getActivity()).saveString("username","0");
-                SharePreferenceSingleton.getInstance(getActivity()).saveString("propic","1");
+                SharePreferenceSingleton.getInstance(getActivity()).saveString("user", "0");
+                SharePreferenceSingleton.getInstance(getActivity()).saveString("username", "0");
+                SharePreferenceSingleton.getInstance(getActivity()).saveString("propic", "1");
 
                 SimpleActivityTransition.goToPreviousActivity(getActivity(), LoginActivity.class);
                 getActivity().finish();
@@ -99,16 +122,85 @@ public class ProfileFragment extends Fragment {
             }
         });
 
+        update.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+
+                Intent intent = new Intent(getActivity(), MainActivity.class)
+                        .putExtra("type", FragmentNode.EDIT_PROFILE);
+                //  .putExtra("receiverImage", currentItem.getProPic());
+                intent.addFlags(Intent.FLAG_ACTIVITY_NO_ANIMATION);
+                intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_NEW_TASK);
+                getActivity().startActivity(intent);
+                getActivity().overridePendingTransition(R.anim.slide_in_from_right, R.anim.fade_out);
+
+            }
+        });
+
         return view;
     }
 
+    private void worksOnAdmin(View view) {
+        changePhoto.setVisibility(View.GONE);
+        final String currentuserName = getActivity().getIntent().getStringExtra("receiver");
+        final String currentuserPass = getActivity().getIntent().getStringExtra("pass");
+        final String tempEMail = currentuserName + "@gmail.com";
+        CardView delete = (CardView) view.findViewById(R.id.delete);
+        delete.setVisibility(View.VISIBLE);
 
-    private void showPictureDialog(){
+        delete.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+
+                final ProgressDialog progressDialog = new ProgressDialog(getActivity());
+                progressDialog.setMessage("Deleting User..");
+                progressDialog.show();
+
+                FirebaseDatabase.getInstance().getReference().child(FirebaseEndPoint.USER_INFO).child(currentuserName).removeValue();
+                final FirebaseAuth mAuth = FirebaseAuth.getInstance();
+                mAuth.signInWithEmailAndPassword(tempEMail, currentuserPass)
+                        .addOnCompleteListener(getActivity(), new OnCompleteListener<AuthResult>() {
+                            @Override
+                            public void onComplete(@NonNull Task<AuthResult> task) {
+                                if (task.isSuccessful()) {
+
+                                    mAuth.getCurrentUser().delete().addOnCompleteListener(new OnCompleteListener<Void>() {
+                                        @Override
+                                        public void onComplete(@NonNull Task<Void> task) {
+
+                                            if (progressDialog!=null){
+                                                progressDialog.dismiss();
+                                            }
+
+                                            SimpleActivityTransition.goToPreviousActivity(getActivity(), AdminActivity.class);
+
+                                        }
+                                    });
+
+                                    if (progressDialog!=null){
+                                        progressDialog.dismiss();
+                                    }
+
+                                } else {
+                                    if (progressDialog!=null){
+                                        progressDialog.dismiss();
+                                    }
+                                }
+                            }
+                        });
+
+            }
+        });
+
+    }
+
+
+    private void showPictureDialog() {
         AlertDialog.Builder pictureDialog = new AlertDialog.Builder(getActivity());
         pictureDialog.setTitle("Select Action");
         String[] pictureDialogItems = {
                 "Select photo from gallery",
-                "Capture photo from camera" };
+                "Capture photo from camera"};
         pictureDialog.setItems(pictureDialogItems,
                 new DialogInterface.OnClickListener() {
                     @Override
@@ -156,7 +248,7 @@ public class ProfileFragment extends Fragment {
 
                 thumbnail = decodeUriToBitmap(getActivity(), contentURI);
 
-                 profileImage.setImageBitmap(thumbnail);
+                profileImage.setImageBitmap(thumbnail);
 
                 imageByte = "2";
 
@@ -179,13 +271,13 @@ public class ProfileFragment extends Fragment {
 
     private void sendToServer(Bitmap thumbnail) {
 
-        if (imageByte.equals("2")){
+        if (imageByte.equals("2")) {
             ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
             thumbnail.compress(Bitmap.CompressFormat.JPEG, 20, byteArrayOutputStream);
-            byte[] byteArray = byteArrayOutputStream .toByteArray();
+            byte[] byteArray = byteArrayOutputStream.toByteArray();
             imageByte = Base64.encodeToString(byteArray, Base64.DEFAULT);
             String username1 = SharePreferenceSingleton.getInstance(getActivity()).getString("username");
-            SharePreferenceSingleton.getInstance(getActivity()).saveString("propic",imageByte);
+            SharePreferenceSingleton.getInstance(getActivity()).saveString("propic", imageByte);
             FirebaseDatabase.getInstance().getReference().child(FirebaseEndPoint.USER_INFO)
                     .child(username1)
                     .child(FirebaseEndPoint.IMAGE)
@@ -209,7 +301,6 @@ public class ProfileFragment extends Fragment {
         }
 
 
-
         return getBitmap;
     }
 
@@ -217,7 +308,7 @@ public class ProfileFragment extends Fragment {
     @Override
     public void onPause() {
 
-        if (valueEventListener!=null && reference!=null){
+        if (valueEventListener != null && reference != null) {
             reference.removeEventListener(valueEventListener);
         }
 
@@ -226,7 +317,7 @@ public class ProfileFragment extends Fragment {
 
     @Override
     public void onStop() {
-        if (valueEventListener!=null && reference!=null){
+        if (valueEventListener != null && reference != null) {
             reference.removeEventListener(valueEventListener);
         }
         super.onStop();
@@ -235,6 +326,12 @@ public class ProfileFragment extends Fragment {
     private void populateData() {
 
         String username1 = SharePreferenceSingleton.getInstance(getActivity()).getString("username");
+        Log.d("username1", username1);
+        if (username1.equals("admin12")) {
+            username1 = getActivity().getIntent().getStringExtra("receiver");
+            Log.d("username1", username1);
+        }
+
         reference = FirebaseDatabase.getInstance().getReference().child("UserData").child(username1);
 
         progressDialog.show();
@@ -254,18 +351,20 @@ public class ProfileFragment extends Fragment {
                 if (dataSnapshot.hasChild(FirebaseEndPoint.LOST)) {
                     String lostCount = String.valueOf(dataSnapshot.child(FirebaseEndPoint.LOST).getChildrenCount());
                     lostNumber.setText(lostCount);
-                }else {
+                } else {
                     lostNumber.setText("0");
                 }
                 if (dataSnapshot.hasChild(FirebaseEndPoint.FOUND)) {
                     String foundCount = String.valueOf(dataSnapshot.child(FirebaseEndPoint.FOUND).getChildrenCount());
                     foundNumber.setText(foundCount);
-                }else {
+                } else {
                     foundNumber.setText("0");
                 }
 
+                UserGeneralInfo userGeneralInfo = dataSnapshot.child(FirebaseEndPoint.USER_GENERAL_INFO).getValue(UserGeneralInfo.class);
 
-                setValue();
+
+                setValue(userGeneralInfo);
                 progressDialog.dismiss();
 
             }
@@ -280,19 +379,19 @@ public class ProfileFragment extends Fragment {
 
     }
 
-    private void setValue() {
+    private void setValue(UserGeneralInfo userGeneralInfo) {
 
-        if (HomeActivity.userGeneralInfo != null) {
-            name.setText(HomeActivity.userGeneralInfo.getName());
-            username.setText(HomeActivity.userGeneralInfo.getUsername());
-            catagory.setText(HomeActivity.userGeneralInfo.getCatagory());
-            email.setText(HomeActivity.userGeneralInfo.getEmail());
-            nsuId.setText(HomeActivity.userGeneralInfo.getNsuId());
-            phone.setText(HomeActivity.userGeneralInfo.getPhone());
-            designation.setText(HomeActivity.userGeneralInfo.getDesignation());
-        }else {
+        if (userGeneralInfo != null) {
+            name.setText(userGeneralInfo.getName());
+            username.setText(userGeneralInfo.getUsername());
+            catagory.setText(userGeneralInfo.getCatagory());
+            email.setText(userGeneralInfo.getEmail());
+            nsuId.setText(userGeneralInfo.getNsuId());
+            phone.setText(userGeneralInfo.getPhone());
+            designation.setText(userGeneralInfo.getDesignation());
+        } else {
 
-            Toast.makeText(getActivity(),"Some data is missing, Please reload",Toast.LENGTH_LONG).show();
+            Toast.makeText(getActivity(), "Some data is missing, Please reload", Toast.LENGTH_LONG).show();
 
         }
 
@@ -318,5 +417,6 @@ public class ProfileFragment extends Fragment {
         progressDialog.setMessage("Please wait..");
         changePhoto = (CardView) view.findViewById(R.id.change_photo);
         logout = (CardView) view.findViewById(R.id.logout);
+        update = (CardView) view.findViewById(R.id.update);
     }
 }
